requirements.txt
# Core Django
Django==4.2.21
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-colorfield==0.14.0
setuptools==80.9.0

# Database
psycopg2-binary==2.9.6
django-postgres-extra==2.0.2

# Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0

# API Documentation
drf-yasg==1.21.7

# Filtering and Search
django-filter==23.5
elasticsearch-dsl==8.11.0

# Image handling
pillow==11.2.1
django-imagekit==5.0.0

# Cache
django-redis==5.4.0
redis==5.0.1

# Background tasks
celery==5.2.7
django-celery-beat==2.5.0
kombu==5.3.4
debugpy==1.8.14

# Phone numbers
django-phonenumber-field==7.2.0
phonenumbers==8.13.25

# Environment variables
python-decouple==3.8
django-environ==0.11.2

# File storage
django-storages==1.14.2
boto3==1.34.0

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipdb==0.13.13

# Production
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Linting and formatting
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Monitoring
django-silk==5.0.4
prometheus-client==0.19.0